advisor-service/requirements.txt
fastapi
uvicorn
langchain
langchain-openai
langchain-community
chromadb
httpx
python-dotenv
pydantic
